+++ benchmark/CMakeLists.txt
@@ -40,7 +40,6 @@ endif()
 target_sources(libfork_benchmark
   PRIVATE
     src/libfork_benchmark/fib/serial.cpp
-    src/libfork_benchmark/fib/serial_return.cpp
 )
 
 # ---- Baremetal ----
