+++ benchmark/src/libfork_benchmark/common.hpp
@@ -1,8 +1,11 @@
 #pragma once
 
+#include <benchmark/benchmark.h>
+
 #include "libfork/__impl/exception.hpp"
 
 import std;
+import libfork;
 
 struct incorrect_result : public std::runtime_error {
   using std::runtime_error::runtime_error;
@@ -19,3 +22,15 @@ inline constexpr unsigned bench_max_threads = 8;
           std::format("{}={} != {}={}", #expected, lf_check_expected_val, #result, lf_check_result_val)));   \
     }                                                                                                        \
   } while (0)
+
+template <lf::scheduler Sch>
+auto make_scheduler(benchmark::State &state) -> Sch {
+  if constexpr (std::constructible_from<Sch, std::size_t>) {
+    return Sch{static_cast<std::size_t>(state.range(1))};
+  } else {
+    return Sch{};
+  }
+}
+
+using mono_busy_pool = lf::mono_busy_pool<lf::geometric_stack<>>;
+using poly_busy_pool = lf::poly_busy_pool<lf::geometric_stack<>>;
