+++ benchmark/src/libfork_benchmark/fib/serial.cpp
@@ -39,7 +39,36 @@ void fib_serial(benchmark::State &state) {
   }
 }
 
+LF_NO_INLINE auto fib(std::int64_t n) -> std::int64_t {
+  if (n < 2) {
+    return n;
+  }
+
+  std::int64_t lhs = fib(n - 1);
+  std::int64_t rhs = fib(n - 2);
+
+  return lhs + rhs;
+}
+
+void fib_serial_return(benchmark::State &state) {
+
+  std::int64_t n = state.range(0);
+  std::int64_t expect = fib_ref(n);
+
+  state.counters["n"] = static_cast<double>(n);
+
+  for (auto _ : state) {
+    benchmark::DoNotOptimize(n);
+    std::int64_t result = fib(n);
+    CHECK_RESULT(result, expect);
+    benchmark::DoNotOptimize(result);
+  }
+}
+
 } // namespace
 
 BENCHMARK(fib_serial)->Name("test/serial/fib")->Arg(fib_test);
 BENCHMARK(fib_serial)->Name("base/serial/fib")->Arg(fib_base);
+
+BENCHMARK(fib_serial_return)->Name("test/serial/fib/return")->Arg(fib_test);
+BENCHMARK(fib_serial_return)->Name("base/serial/fib/return")->Arg(fib_base);
